refactor: 쿠폰 할인타입 상수화
